use new link predicates / sources & remove old polling code
