+++ src/store/data/actions/index.ts
@@ -1,23 +1,19 @@
 import createChannel from "./createChannel";
 import createCommunity from "./createCommunity";
-import getNeighbourhoodChannels from "./getNeighbourhoodChannels";
-import getNeighbourhoodMetadata from "./getNeighbourhoodMetadata";
-import getNeighbourhoodMembers from "./getNeighbourhoodMembers";
 import joinCommunity from "./joinCommunity";
 import updateCommunity from "./updateCommunity";
 import showMessageNotification from "./showMessageNotification";
 import fetchNeighbourhoodMembers from "./fetchNeighbourhoodMembers";
 import fetchNeighbourhoodMetadata from "./fetchNeighbourhoodMetadata";
+import fetchNeighbourhoodChannels from "./fetchNeighbourhoodChannels";
 
 export default {
   createChannel,
   createCommunity,
-  getNeighbourhoodChannels,
-  getNeighbourhoodMetadata,
-  getNeighbourhoodMembers,
   joinCommunity,
   updateCommunity,
   showMessageNotification,
   fetchNeighbourhoodMembers,
   fetchNeighbourhoodMetadata,
+  fetchNeighbourhoodChannels
 };
