+++ vite.config.js
@@ -21,4 +21,14 @@ export default {
       "@": path.resolve(__dirname, "./src"),
     },
   },
+  server: {
+    port: 3030,
+    proxy: {
+      "/dist/main.js": {
+        target: "http://127.0.0.1:8080",
+        changeOrigin: true,
+        secure: false,
+      },
+    },
+  },
 };
